Move filter-relevant javascript to seperate file
